+++ docs/fr_FR/Changelog.rst
@@ -4,7 +4,8 @@ ChangeLog
 - Interne: Cmd: Suppression 'Network_Address_request' => 'getNwkAddress'.
 - Page équipment/avancé: Ajout support 'Network_Address_request'.
 - Interne: Cmd: 'cmd-Private' revisité.
-- Forcer un modèle: Corrections en cours.
+- Modèle forcé: Corrections.
+- Profalux: Cmd: Corrections.
 
 240209-BETA-1
 -------------